Small error at: FAILED: [code=1] CMakeFiles/zithc_objects.dir/src/lexer/lexer.cpp.o                 /usr/local/bin/zig c++ -target x86_64-linux-musl -DZITH_LIB_DIR=\"/home/runner/work/Zith/Zith/lib\" -DZITH_VERSION=\"v0.0.3.1\" -I/home/runner/work/Zith/Zith/src -I/home/runner/work/Zith/Zith/build/_deps/mio-src/include -isystem /home/runner/work/Zith/Zith/build/_deps/tomlplusplus-src/include -O3 -DNDEBUG -std=gnu++2b -MD -MT CMakeFiles/zithc_objects.dir/src/lexer/lexer.cpp.o -MF CMakeFiles/zithc_objects.dir/src/lexer/lexer.cpp.o.d -o CMakeFiles/zithc_objects.dir/src/lexer/lexer.cpp.o -c /home/runner/work/Zith/Zith/src/lexer/lexer.cpp                 /home/runner/work/Zith/Zith/src/lexer/lexer.cpp:88:26: error: no matching constructor for initialization of 'std::string_view' (aka 'basic_string_view<char>')                         std::string_view rest{now, end};                                          ^   ~~~~~~~~~~                 /opt/zig-linux-x86_64-0.9.1/lib/libcxx/include/string_view:265:5: note: candidate constructor not viable: no known conversion from 'const char *' to 'std::basic_string_view::size_type' (aka 'unsigned long') for 2nd argument; dereference the argument with *                     basic_string_view(const _CharT* __s, size_type __len) _NOEXCEPT                     ^                 /opt/zig-linux-x86_64-0.9.1/lib/libcxx/include/string_view:259:5: note: candidate constructor not viable: requires 1 argument, but 2 were provided                     basic_string_view(const basic_string_view&) _NOEXCEPT = default;                     ^                 /opt/zig-linux-x86_64-0.9.1/lib/libcxx/include/string_view:274:5: note: candidate constructor not viable: requires single argument '__s', but 2 arguments were provided                     basic_string_view(const _CharT* __s)                     ^                 /opt/zig-linux-x86_64-0.9.1/lib/libcxx/include/string_view:278:5: note: candidate constructor not viable: requires 1 argument, but 2 were provided                     basic_string_view(nullptr_t) = delete;                     ^                 /opt/zig-linux-x86_64-0.9.1/lib/libcxx/include/string_view:256:5: note: candidate constructor not viable: requires 0 arguments, but 2 were provided                     basic_string_view() _NOEXCEPT : __data (nullptr), __size(0) {}                     ^                 1 error generated.         Zig small fix

+++ src/cli/options.cpp
@@ -305,7 +305,7 @@ Options parseArgs(int argc, char **argv) {
             for (const char *p = val; ; ++p) {
                 if (*p == ',' || *p == '\0') {
                     if (p > start) {
-                        opts.include_dirs.emplace(start, p - start);
+                        opts.include_dirs.emplace(start, static_cast<size_t>(p - start));
                     }
                     if (*p == '\0')
                         break;
